Split Color to its own file
